nifi_pipelines/ai-tutor/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.3.1
boto3==1.38.35
botocore==1.38.35
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.3.0
frozenlist==1.7.0
google-auth==2.40.3
google-genai==1.19.0
grpcio==1.73.0
grpcio-health-checking==1.73.0
grpcio-tools==1.73.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
multidict==6.4.4
openai==1.86.0
packaging==25.0
propcache==0.3.2
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==1.26.20
validators==0.34.0
weaviate-client==4.15.0
websockets==15.0.1
yarl==1.20.1
